requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dataclasses-json==0.6.7
ddgs==9.9.1
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
fake-useragent==2.2.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.7
langchain-anthropic==1.0.4
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lxml==6.0.2
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.8.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
primp==0.15.0
propcache==0.4.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
